requirements.txt
annotated-types==0.7.0
anthropic==0.31.0
anyio==4.4.0
APScheduler==3.11.0
attrs==25.3.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
filelock==3.15.4
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==2.1.5
mistune==3.1.3
numpy==1.26.4
openai==1.35.3
opencv-python==4.10.0.82
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypdfium2==4.30.0
pytesseract==0.3.10
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
